fix examples, consistent naming
