+++ examples/sympy_backends.py
@@ -10,27 +10,23 @@ amplitudes.
 """
 
 from qiskit_addon_sympy import SympyProvider
-# The following import will be replaced by an import of register
-# once register is written
-from qiskit.wrapper._wrapper import _DEFAULT_PROVIDER
-from qiskit import execute, load_qasm_file
+from qiskit import register, load_qasm_file, execute
 
 
 def use_sympy_backends():
     """ Usage examples for the Sympy simulators """
 
-    # The following lines will be replaced by a usage of register
-    provider = SympyProvider()
-    _DEFAULT_PROVIDER.add_provider(provider)
+    # register with the SympyProvider to get access to its simulators
+    register(provider_class=SympyProvider)
     q_circuit = load_qasm_file('simple.qasm')
 
     # sympy statevector simulator
-    result = execute(q_circuit, backend='local_statevector_simulator_sympy').result()
+    result = execute(q_circuit, backend='sympy_statevector_simulator').result()
     print("final quantum amplitude vector: ")
     print(result.get_statevector(q_circuit))
 
     # sympy unitary simulator
-    result = execute([q_circuit], backend='local_unitary_simulator_sympy').result()
+    result = execute([q_circuit], backend='sympy_unitary_simulator').result()
     print("\nunitary matrix of the circuit: ")
     print(result.get_unitary(q_circuit))
