+++ test/test_sympy_statevector_simulator.py
@@ -18,8 +18,8 @@ from qiskit import (load_qasm_file, execute, QuantumRegister,
 from qiskit_addon_sympy import SympyProvider
 
 
-class UnitarySimulatorSympyTest(QiskitSympyTestCase):
-    """Test local unitary simulator sympy."""
+class SympyStatevectorSimulatorTest(QiskitSympyTestCase):
+    """Test local statevector simulator."""
 
     def setUp(self):
         wrapper.register(provider_class=SympyProvider)
@@ -27,28 +27,16 @@ class UnitarySimulatorSympyTest(QiskitSympyTestCase):
         self.qasm_filename = self._get_resource_path('simple.qasm')
         self.q_circuit = load_qasm_file(self.qasm_filename)
 
-    def test_unitary_simulator(self):
-        """test generation of circuit unitary"""
+    def test_sympy_statevector_simulator(self):
+        """Test final state vector for single circuit run."""
+        result = execute(self.q_circuit, backend='sympy_statevector_simulator').result()
+        actual = result.get_statevector(self.q_circuit)
 
-        result = execute(self.q_circuit, backend='sympy_unitary_simulator').result()
-        actual = result.get_unitary(self.q_circuit)
-
-        self.assertEqual(actual[0][0], sqrt(2)/2)
-        self.assertEqual(actual[0][1], sqrt(2)/2)
-        self.assertEqual(actual[0][2], 0)
-        self.assertEqual(actual[0][3], 0)
-        self.assertEqual(actual[1][0], 0)
-        self.assertEqual(actual[1][1], 0)
-        self.assertEqual(actual[1][2], sqrt(2)/2)
-        self.assertEqual(actual[1][3], -sqrt(2)/2)
-        self.assertEqual(actual[2][0], 0)
-        self.assertEqual(actual[2][1], 0)
-        self.assertEqual(actual[2][2], sqrt(2)/2)
-        self.assertEqual(actual[2][3], sqrt(2)/2)
-        self.assertEqual(actual[3][0], sqrt(2)/2)
-        self.assertEqual(actual[3][1], -sqrt(2)/2)
-        self.assertEqual(actual[3][2], 0)
-        self.assertEqual(actual[3][3], 0)
+        self.assertEqual(result.get_status(), 'COMPLETED')
+        self.assertEqual(actual[0], sqrt(2)/2)
+        self.assertEqual(actual[1], 0)
+        self.assertEqual(actual[2], 0)
+        self.assertEqual(actual[3], sqrt(2)/2)
 
 
 class TestQobj(QiskitSympyTestCase):
@@ -66,8 +54,8 @@ class TestQobj(QiskitSympyTestCase):
         self.cr_name = cr.name
         self.circuits = [qc]
 
-    def test_qobj_unitary_simulator_sympy(self):
-        qobj = wrapper.compile(self.circuits, backend='sympy_unitary_simulator')
+    def test_qobj_sympy_statevector_simulator(self):
+        qobj = wrapper.compile(self.circuits, backend='sympy_statevector_simulator')
         cc = qobj['circuits'][0]['compiled_circuit']
         ccq = qobj['circuits'][0]['compiled_circuit_qasm']
         self.assertIn(self.qr_name, map(lambda x: x[0], cc['header']['qubit_labels']))
